requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.4.1
djangorestframework==3.15.2
idna==3.10
packaging==24.2
pillow==11.1.0
psycopg2==2.9.6
psycopg2-binary==2.8.6
PyJWT==2.9.0
pytz==2025.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
whitenoise==6.9.0
